refact: matchLottos 메서드에서 WinnerNumbers 객체와 Lotto가 서로 양방향 의존인데, 단방향으로 처리할 것
