+++ README.md
@@ -157,7 +157,7 @@
 
 ### 기능목록
 
-- [ ] matchLottos 메서드에서 WinnerNumbers 객체와 Lotto가 서로 양방향 의존인데, 단방향으로 처리할 것
+- [X] ~~matchLottos 메서드에서 WinnerNumbers 객체와 Lotto가 서로 양방향 의존인데, 단방향으로 처리할 것~~
 - [ ] 사용자가 수동으로 추첨 번호를 입력할 수 있도록 해야 한다.
   - 입력한 금액, 자동 생성 숫자, 수동 생성 번호를 입력하도록 한다.
 - [ ] 로또 구입 금액을 입력하면, 구입 금액에 해당하는 로또를 발급한다.
